Move LLM class method tests into llm_test file
These tests do not require Net::HTTP instrumentation.
Now, the location of the tests more closely align with
where they are defined.
